# Speed up parsing simple unit strings with `Unit`