--- a/docs/changes/units/17004.perf.rst
+++ b/docs/changes/units/17004.perf.rst
@@ -1,0 +1,2 @@
+Parsing strings representing non-composite units with ``Unit`` is now up to 25%
+faster.