@eerovaher 2024-09-13
I don't know what the definition of "significant" is in our benchmarks, but if you dig through the logs you will see that the difference between the relevant timings is much larger than their uncertainty, i.e. the change is significant: 

https://github.com/astropy/astropy/actions/runs/10836066679/job/30068962593?pr=17004#step:7:636

https://github.com/astropy/astropy/actions/runs/10836066679/job/30068962593?pr=17004#step:7:1143