@eerovaher 2024-09-13
> ```
>             # seems more logical to switch order
>             if isinstance(s, bytes):
>                 s = s.decode("ascii")
> ```
There was no significant speedup from placing this check after the shortcut.

> ``` 
>             # might as well speed this up too.
>             f = unit_format.Generic if format is None else unit_format.get_format(format)
> ```
Changing this part had no noticeable effect on performance, but getting rid of the redundant `if format is None` check does simplify the code.

> ```
>             # For Generic, try to short-circuit simple units.
>             if (f is unit_format.Generic
>                     and (unit := get_current_unit_registry().registry.get(s))):
>                 return unit
> ```
Why wouldn't we want to speed up the other unit formats too?