@mhvk 2024-09-13
The main benefit I saw of my suggestion is that it does not rely on private methods at all. But it should also be (marginally) faster for the by-far most common case of `format=None`, both for simple units and for complex ones (latter because of not having to deal with a failed `try/except`). I don't think there is much need to worry about parsing in other formats - that is almost exclusively done during I/O where the reading of associated data sets dominates.