@pllim 2024-09-13
Re: https://github.com/astropy/astropy/pull/17004#issuecomment-2348878833 -- Maybe @MridulS knows since he implemented the infrastructure.