@mhvk 2024-09-13
> The following patch causes multiple test failures:

The reason is that your other change is interfering; if starting from your PR, I remove the try/except you added in `core.py`, one also gets failures.