@eerovaher 2024-09-13
My original suggestion was to move a shortcut from `Generic._do_parse()` to `_UnitMetaClass`. Are you suggesting we should keep the current shortcut in `Generic._do_parse()` and implement an additional new shortcut in `_UnitMetaClass`?