@mhvk 2024-09-14
I'm actually not sure what is the best way right now. Certainly, the failure I mentioned above shows that current implementation is a bit broken - see #17010. I'll see if I can find some time over the weekend to solve that - that may also make clear what the best path forward is.