@mhvk 2024-09-15
I found that removing the short-circuit showed quite a few bugs - those also show up if one tries to create composite units, and are fixed in #17011. I think this in itself makes the case stronger not to try too much short-circuiting in the parsers - they should just parse! So, I had a trial for moving it up in #17012.